tests/wpt/web-platform-tests/css/css-position/parsing/left-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Positioned Layout Module Level 3: getComputedValue().left</title>
<link rel="help" href="https://drafts.csswg.org/css-position/#propdef-left">
<meta name="assert" content="left lengths are made absolute.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<style>
  #target {
    font-size: 40px;
  }
</style>
<script>
test_computed_value("left", "auto");

test_computed_value("left", "calc(10px + 0.5em)", "30px");
test_computed_value("left", "calc(10px - 0.5em)", "-10px");
test_computed_value("left", "-40%");
test_computed_value("left", "calc(50% + 60px)");
</script>
</body>
</html>
